Revert "Refactor the Forger class and move the genesis block related logic outside the Forger class"
