+++ src/tinychain.py
@@ -14,8 +14,6 @@ from vm import TinyVMEngine
 from wallet import Wallet
 from parameters import HTTP_PORT, MAX_TX_POOL, PEER_URIS
 from peer_communication import broadcast_block_header, receive_block_header, broadcast_transaction
-from forger import Forger
-from genesis import genesis_procedure
 
 TINYCOIN = 1000000000000000000
 TINYCHAIN_UNIT = 'tatoshi'
@@ -41,6 +39,239 @@ class TransactionPool:
     def is_empty(self):
         return not self.transactions
 
+
+class Forger:
+    def __init__(self, transactionpool, storage_engine, validation_engine, wallet):
+        self.transactionpool = transactionpool
+        self.storage_engine = storage_engine
+        self.validation_engine = validation_engine
+
+        self.wallet = wallet
+        self.validator = self.proposer = self.wallet.get_address()
+        self.sign = self.wallet.sign_message
+
+        self.production_enabled = True
+
+        self.in_memory_blocks = {}  # P7e15
+        self.in_memory_block_headers = {}  # P7e15
+        self.current_proposer_index = 0  # Initialize the current proposer index
+
+    @staticmethod
+    def generate_block_hash(merkle_root, timestamp, state_root, previous_block_hash):
+        values = [merkle_root, str(timestamp), str(state_root), previous_block_hash]
+        concatenated_string = f"{''.join(values)}".encode()
+        return blake3.blake3(concatenated_string).hexdigest()
+
+    @staticmethod
+    def compute_merkle_root(transaction_hashes):
+        if len(transaction_hashes) == 0:
+            return blake3.blake3(b'').hexdigest()
+
+        while len(transaction_hashes) > 1:
+            if len(transaction_hashes) % 2 != 0:
+                transaction_hashes.append(transaction_hashes[-1])
+
+            if isinstance(transaction_hashes[0], bytes):
+                transaction_hashes = [blake3.blake3(transaction_hashes[i] + transaction_hashes[i + 1]).digest() for i in range(0, len(transaction_hashes), 2)]
+            elif isinstance(transaction_hashes[0], str):
+                transaction_hashes = [blake3.blake3(transaction_hashes[i].encode() + transaction_hashes[i + 1].encode()).digest() for i in range(0, len(transaction_hashes), 2)]
+            else:
+                raise TypeError("Unsupported data type in transaction_hashes")
+            
+        if isinstance(transaction_hashes[0], str):
+            transaction_hashes[0] = transaction_hashes[0].encode('utf-8')
+
+        return blake3.blake3(transaction_hashes[0]).hexdigest()
+
+    def toggle_production(self):
+        self.production_enabled = not self.production_enabled
+
+    def fetch_current_validator_set(self):
+        staking_contract_state = self.storage_engine.fetch_contract_state("7374616b696e67")
+        if staking_contract_state:
+            return sorted(staking_contract_state.keys(), key=lambda k: staking_contract_state[k]['index'])
+        return []
+
+    def select_proposer(self):
+        validator_set = self.fetch_current_validator_set()
+        if validator_set:
+            proposer = validator_set[self.current_proposer_index]
+            self.current_proposer_index = (self.current_proposer_index + 1) % len(validator_set)
+            return proposer
+        return None
+
+    def forge_new_block(self, replay=True, block_header=None, is_genesis=False):
+        if not self.production_enabled:
+            return "Forging is disabled"
+
+        transactions_to_forge = []
+
+        if replay:
+            for transaction_hash in block_header.transaction_hashes:
+                transaction = self.transactionpool.get_transaction_by_hash(transaction_hash)
+                if transaction is not None:
+                    transactions_to_forge.append(transaction)
+                else:
+                    logging.info(f"Transaction {transaction_hash} not found in pool, requesting transaction from peers...")
+                    # todo: request transaction from connected peers (or the block producer?)
+        else:
+            transactions_to_forge = self.transactionpool.get_transactions()
+
+        if not is_genesis:
+            valid_transactions_to_forge = [t for t in transactions_to_forge if self.validation_engine.validate_transaction(t)]
+        else:
+            valid_transactions_to_forge = transactions_to_forge
+
+        # Get the previous block for validation
+        if is_genesis is False:
+            previous_block_header = self.storage_engine.fetch_last_block_header()
+            previous_block_hash = previous_block_header.block_hash
+            height = previous_block_header.height + 1
+            current_state = self.storage_engine.fetch_state(previous_block_header.state_root)
+        else:
+            height = 0
+            current_state = {}
+
+        tvm_engine = TinyVMEngine(current_state)
+
+        if not replay:
+            if not is_genesis:
+                ### NEW BLOCK PROPOSAL CASE ###
+                self.proposer = self.wallet.get_address()
+                timestamp = int(time.time())
+                state_root, new_state = tvm_engine.exec(valid_transactions_to_forge, self.proposer)
+                transaction_hashes = [t.to_dict()['transaction_hash'] for t in valid_transactions_to_forge]
+                merkle_root = self.compute_merkle_root(transaction_hashes)
+                block_hash = self.generate_block_hash(merkle_root, timestamp, state_root, previous_block_hash)
+                signature = self.sign(block_hash)
+                validator_index = self.get_validator_index(self.proposer)
+                signatures = [Signature(self.proposer, timestamp, signature, validator_index)]
+            else:
+                ### GENESIS BLOCK CASE ###
+                self.proposer = "genesis"
+                timestamp = int(19191919)
+                state_root, new_state = tvm_engine.exec(transactions_to_forge, self.proposer)
+                transaction_hashes = [t.to_dict()['transaction_hash'] for t in transactions_to_forge]
+                merkle_root = self.compute_merkle_root(transaction_hashes)
+                previous_block_hash = "00000000"
+                block_hash = self.generate_block_hash(merkle_root, timestamp, state_root, previous_block_hash)
+                signature = "genesis_signature"
+                validator_index = -1
+                signatures = [Signature(self.proposer, timestamp, signature, validator_index)]
+
+            block_header = BlockHeader(
+                block_hash,
+                height,
+                timestamp,
+                previous_block_hash,
+                merkle_root,
+                state_root,
+                self.proposer,
+                signatures,
+                transaction_hashes
+            )
+        else:
+            ### BLOCK REPLAY CASE ###
+            state_root, new_state = tvm_engine.exec(valid_transactions_to_forge, block_header.proposer)
+            if state_root == block_header.state_root:
+                transaction_hashes = [t.to_dict()['transaction_hash'] for t in block_header.transactions]
+                computed_merkle_root = self.compute_merkle_root(transaction_hashes)
+                if computed_merkle_root == block_header.merkle_root:
+                    signature = self.wallet.sign_message(block_hash)
+                    validator_index = self.get_validator_index(self.validator)
+                    signatures = block_header.signatures
+                    logging.info("Block signatures: %s", signatures)
+
+                    if isinstance(signatures, list) and all(isinstance(sig, Signature) for sig in signatures):
+                        signatures.append(Signature(self.validator, int(time.time()), signature, validator_index))
+                    else:
+                        signatures = [Signature.from_dict(sig) for sig in signatures]
+                        signatures.append(Signature(self.validator, int(time.time()), signature, validator_index))
+
+                    block_header = BlockHeader(
+                        block_header.block_hash,
+                        block_header.height,
+                        block_header.timestamp,
+                        block_header.previous_block_hash,
+                        block_header.merkle_root,
+                        block_header.state_root,
+                        block_header.proposer,
+                        signatures,
+                        block_header.transaction_hashes
+                    )
+                    logging.info("Replay successful for block %s", block_header.block_hash)
+                else:
+                    logging.error("Replay failed for block %s (Merkle root mismatch)", block_header.block_hash)
+                    return False
+            else:
+                logging.error("Replay failed for block %s (State root mismatch)", block_header.block_hash)
+                return False
+        
+        if not is_genesis:
+            block = Block(block_header, valid_transactions_to_forge)
+        else:
+            block = Block(block_header, transactions_to_forge)
+
+        if not is_genesis:
+            self.in_memory_blocks[block.header.block_hash] = block
+            self.in_memory_block_headers[block.header.block_hash] = block_header
+
+            broadcast_block_header(block_header)
+
+            if self.has_enough_signatures(block_header):
+
+                self.storage_engine.store_block(block)
+                self.storage_engine.store_block_header(block_header)
+                self.storage_engine.store_state(block.header.state_root, new_state)
+                return True
+            else:
+                # Drop the block if not enough signatures at block_timeout
+                del self.in_memory_blocks[block.header.block_hash]
+                del self.in_memory_block_headers[block.header.block_hash]
+                return False
+        else:
+            self.storage_engine.store_block(block)
+            self.storage_engine.store_block_header(block_header)
+            self.storage_engine.store_state(block.header.state_root, new_state)
+            return True
+
+    def has_enough_signatures(self, block_header):  # P66ad
+        # Logic to check if 2/3 validators have signed
+        return True
+
+    def get_validator_index(self, validator_address):
+        staking_contract_state = self.storage_engine.fetch_contract_state("7374616b696e67")
+        if staking_contract_state and validator_address in staking_contract_state:
+            return staking_contract_state[validator_address]['index']
+        return -1
+
+def genesis_procedure():
+    genesis_addresses = [
+        "7ff08d4ddd1be1305e77db4064bb71f2c0872599334db03fc36d8cab3fb349c4e1dfb6262bd3e118f37aa8d19827f0aa56bf9052bb9c5b6e16e0679706124e38",
+        "336bf5193983cb2c287f7695343730ccf5f0a88da961459e996b1d4cc07480f1649a682e1194db36bd9957e6cb6cfb4c2185306f31da9723bc84b03e4116ce57",
+        "448769135d4bef3c2b88f81829dc73433fd05882082bb67c4962ba9ef2acf6967083d7f142758150f6f01a87e7a18aa568400b5a05b2b24ccce415d764fed636",
+        "f0689410ce583320822fb2761600b1786860711ec751fae27d2341ea1befbcc0a3c1afde193e9d2d5d9a9f5a79c7f18efa49e4f5905767c4d518b9c78850305f",
+        "ce8958afa7c8ac763308705cda306d21119f5401d08d249b1cdcbd6306faba91c6f31a210c7827a494891d17d357dfa755a6e4ae5dd09ffa3803fbafe58c027c",
+        "de280016480735a12e975d4c869808bdee564d544e762020446462261d704d8d4a8cab311c6447e3299108b7317db9d5316057c0c0cf7379dda2cd12c20d7660"
+    ]
+    staking_contract_address = "7374616b696e67"
+    genesis_transactions = [
+        Transaction("genesis", genesis_addresses[0], 10000*TINYCOIN, 120, 0, "consensus", ""),
+        Transaction(genesis_addresses[0], staking_contract_address, 1000*TINYCOIN, 110, 0, "genesis_signature_0", "stake"),
+        Transaction("genesis", genesis_addresses[1], 10000*TINYCOIN, 100, 1, "consensus", ""),
+        Transaction(genesis_addresses[1], staking_contract_address, 1000*TINYCOIN, 90, 0, "genesis_signature_1", "stake"),
+        Transaction("genesis", genesis_addresses[2], 10000*TINYCOIN, 80, 2, "consensus", ""),
+        Transaction(genesis_addresses[2], staking_contract_address, 1000*TINYCOIN, 70, 0, "genesis_signature_2", "stake"),
+        Transaction("genesis", genesis_addresses[3], 10000*TINYCOIN, 60, 3, "consensus", ""),
+        Transaction(genesis_addresses[3], staking_contract_address, 1000*TINYCOIN, 50, 0, "genesis_signature_3", "stake"),
+        Transaction("genesis", genesis_addresses[4], 10000*TINYCOIN, 40, 4, "consensus", ""),
+        Transaction(genesis_addresses[4], staking_contract_address, 1000*TINYCOIN, 30, 0, "genesis_signature_4", "stake"),
+        Transaction("genesis", genesis_addresses[5], 10000*TINYCOIN, 20, 5, "consensus", ""),
+        Transaction(genesis_addresses[5], staking_contract_address, 1000*TINYCOIN, 10, 0, "genesis_signature_5", "stake")
+    ]
+    for transaction in genesis_transactions:
+        transactionpool.add_transaction(transaction)
+    forger.forge_new_block(False, None, True)
     
 class StorageEngine:
     def __init__(self, transactionpool):
